# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FilmCraft - Full filmmaking production suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List characters for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterBody"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /projects/{projectId}/characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharacterBody"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── BEAT SHEETS ────────────────────────────────────────────────────────────
...
  /projects/{projectId}/mindmap:
    get:
      operationId: getMindMap
      tags: [mindmap]
      summary: Get mind map nodes for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mind map nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MindMapNode"
    post:
      operationId: createMindMapNode
      tags: [mindmap]
      summary: Create a mind map node
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMindMapNodeBody"
      responses:
        "201":
          description: Created node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MindMapNode"
  /projects/{projectId}/mindmap/{id}:
    patch:
      operationId: updateMindMapNode
      tags: [mindmap]
      summary: Update a mind map node
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMindMapNodeBody"
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MindMapNode"
    delete:
      operationId: deleteMindMapNode
      tags: [mindmap]
      summary: Delete a mind map node
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SCENES ──────────────────────────────────────────────────────────────────
...
  /projects/{projectId}/crew:
    get:
      operationId: listCrewMembers
      tags: [crew]
      summary: List crew members for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crew members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrewMember"
    post:
      operationId: createCrewMember
      tags: [crew]
      summary: Add a crew member
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrewMemberBody"
      responses:
        "201":
          description: Created crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
  /projects/{projectId}/crew/{id}:
    patch:
      operationId: updateCrewMember
      tags: [crew]
      summary: Update a crew member
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCrewMemberBody"
      responses:
        "200":
          description: Updated crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
    delete:
      operationId: deleteCrewMember
      tags: [crew]
      summary: Delete a crew member
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SHOT LIST ───────────────────────────────────────────────────────────────
...
  /projects/{projectId}/lighting:
    get:
      operationId: listLightingDiagrams
      tags: [lighting]
      summary: List lighting diagrams for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lighting diagrams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LightingDiagram"
    post:
      operationId: createLightingDiagram
      tags: [lighting]
      summary: Create a lighting diagram
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLightingDiagramBody"
      responses:
        "201":
          description: Created lighting diagram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LightingDiagram"
  /projects/{projectId}/lighting/{id}:
    get:
      operationId: getLightingDiagram
      tags: [lighting]
      summary: Get a lighting diagram with all elements
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lighting diagram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LightingDiagram"
    patch:
      operationId: updateLightingDiagram
      tags: [lighting]
      summary: Update a lighting diagram (including its elements JSON)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLightingDiagramBody"
      responses:
        "200":
          description: Updated lighting diagram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LightingDiagram"
    delete:
      operationId: deleteLightingDiagram
      tags: [lighting]
      summary: Delete a lighting diagram
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTION PACKETS ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Character:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        role:
          type: string
          nullable: true
          description: "protagonist | antagonist | supporting | minor"
        age:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        backstory:
          type: string
          nullable: true
        motivation:
          type: string
          nullable: true
        arc:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - name
        - createdAt
    CreateCharacterBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          nullable: true
        age:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        backstory:
          type: string
          nullable: true
        motivation:
          type: string
          nullable: true
        arc:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
    UpdateCharacterBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          nullable: true
        age:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        backstory:
          type: string
          nullable: true
        motivation:
          type: string
          nullable: true
        arc:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── BEAT ─────────────────────────────────────────────────────────────────
...
    MindMapNode:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        label:
          type: string
        nodeType:
          type: string
          nullable: true
          description: "root | theme | character | location | conflict | idea"
        parentId:
          type: integer
          nullable: true
        positionX:
          type: number
          nullable: true
        positionY:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - label
        - createdAt
    CreateMindMapNodeBody:
      type: object
      properties:
        label:
          type: string
        nodeType:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        positionX:
          type: number
          nullable: true
        positionY:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - label
    UpdateMindMapNodeBody:
      type: object
      properties:
        label:
          type: string
        nodeType:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        positionX:
          type: number
          nullable: true
        positionY:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── SCENE ─────────────────────────────────────────────────────────────────
...
    CrewMember:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        department:
          type: string
          nullable: true
          description: "direction | camera | lighting | sound | art | wardrobe | makeup | production | post"
        title:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        availability:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - name
        - createdAt
    CreateCrewMemberBody:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        availability:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
    UpdateCrewMemberBody:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        availability:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── SHOT ──────────────────────────────────────────────────────────────────
...
    LightingDiagram:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        sceneId:
          type: integer
          nullable: true
        canvasData:
          type: string
          nullable: true
          description: "JSON blob storing all diagram elements (lights, cameras, subjects, walls, furniture)"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - name
        - createdAt
        - updatedAt
    CreateLightingDiagramBody:
      type: object
      properties:
        name:
          type: string
        sceneId:
          type: integer
          nullable: true
        canvasData:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
    UpdateLightingDiagramBody:
      type: object
      properties:
        name:
          type: string
        sceneId:
          type: integer
          nullable: true
        canvasData:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── PRODUCTION PACKET ─────────────────────────────────────────────────────
...